Added @Data annotation to order entity

+++ orders-service/src/main/java/com/elbarqy/appstore/orderservice/command/models/CreateOrderCommand.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Data;
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class CreateOrderCommand {
@@ -11,7 +13,7 @@ public class CreateOrderCommand {
     private final String userID;
     public final String orderID;
     private final String productID;
-    private final int quantity;
+    private final BigDecimal quantity;
     private final String addressID;
     private final OrderStatus orderStatus;
 }
