+++ orders-service/src/main/java/com/elbarqy/appstore/orderservice/command/models/CreateOrderRest.java
@@ -4,13 +4,14 @@ import lombok.Data;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
+import java.math.BigDecimal;
 
 @Data
 public class CreateOrderRest {
     @NotBlank
     private String productID;
     @Min(value = 1)
-    private int quantity;
+    private BigDecimal quantity;
     @NotBlank
     private String addressID;
 }
