+++ orders-service/src/main/java/com/elbarqy/appstore/orderservice/command/models/OrderCreatedEvent.java
@@ -2,12 +2,14 @@ package com.elbarqy.appstore.orderservice.command.models;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderCreatedEvent {
     private String userID;
     public String orderID;
     private String productID;
-    private int quantity;
+    private BigDecimal quantity;
     private String addressID;
     private OrderStatus orderStatus;
 }
